+++ compatibility-kit/ruby/features/minimal/minimal.feature.rb
@@ -1,2 +1,0 @@
-Given('I have {int} cukes in my belly') do |cuke_count|
-end