+++ compatibility-kit/ruby/features/undefined/undefined.feature.rb
@@ -1,7 +1,0 @@
-Given('an implemented step') do
-  # no-op
-end
-
-Given('a step that we expect to be skipped') do
-  # no-op
-end