+++ fake-cucumber/javascript/src/IParameterTypeDefinition.ts
@@ -1,8 +1,0 @@
-export default interface IParameterTypeDefinition {
-  name: string
-  regexp: RegExp | readonly RegExp[] | string | readonly string[]
-  type?: any
-  transformer?: (...args: readonly string[]) => any
-  preferForRegexpMatch?: boolean
-  useForSnippets?: boolean
-}