+++ json-to-messages/python-testdata/features/hooks/steps/hooks.py
@@ -1,6 +1,0 @@
-from behave import *
-
-@given('a passed step')
-def step_impl(context):
-    pass
-